+++ core/db/Redis.php
@@ -18,7 +18,8 @@ use core\lib\Config;
  */
 class Redis extends Driver {
 
-    protected $options = [];
+    protected $options   = [];
+    protected $is_active = null;
     /**
      * @var Redis
      */
@@ -37,16 +38,23 @@ class Redis extends Driver {
         if (!$this->handler or !$this->handler instanceof \Redis) {
             $this->handler = new \Redis();
         }
-        if ($this->options['persistent']) {
-            $this->handler->pconnect($this->options['host'], $this->options['port'], $this->options['timeout'], 'persistent_id_' . $this->options['select']);
-        } else {
-            $this->handler->connect($this->options['host'], $this->options['port'], $this->options['timeout']);
-        }
-        if ('' != $this->options['password']) {
-            $this->handler->auth($this->options['password']);
-        }
-        if (0 != $this->options['select']) {
-            $this->handler->select($this->options['select']);
+        try{
+            if ($this->options['persistent']) {
+                $this->handler->pconnect($this->options['host'], $this->options['port'], $this->options['timeout'], 'persistent_id_' . $this->options['select']);
+            } else {
+                $this->handler->connect($this->options['host'], $this->options['port'], $this->options['timeout']);
+            }
+
+            if ('' != $this->options['password']) {
+                $this->handler->auth($this->options['password']);
+            }
+
+            if (0 != $this->options['select']) {
+                $this->handler->select($this->options['select']);
+            }
+        }catch (\Exception $e){
+            $this->is_active = false;
+            log_add('redis server exception','REDIS',__METHOD__);
         }
     }
 
@@ -61,6 +69,7 @@ class Redis extends Driver {
             }
             return false;
         }
+        return true;
     }
 
     /**
@@ -81,8 +90,7 @@ class Redis extends Driver {
      * @return bool
      */
     public function has($name) {
-        $this->_ext();
-        $this->call();
+        if(!$this->_ext() or !$this->call()) return false;
         return $this->handler->exists($this->getCacheKey($name));
     }
 
@@ -91,8 +99,7 @@ class Redis extends Driver {
      * @return array|bool
      */
     public function keys($name){
-        $this->_ext();
-        $this->call();
+        if(!$this->_ext() or !$this->call()) return false;
         return $this->handler->keys($this->getCacheKey($name));
     }
 
@@ -116,11 +123,10 @@ class Redis extends Driver {
 
     /**
      * @param $name
-     * @return array
+     * @return array|bool
      */
     public function hGetAll($name){
-        $this->_ext();
-        $this->call();
+        if(!$this->_ext() or !$this->call()) return false;
         $h = $this->handler->hGetAll($this->getCacheKey($name));
         if($h){
             foreach ($h as &$value){
@@ -132,17 +138,17 @@ class Redis extends Driver {
 
     /**
      * @param $key
-     * @return array
+     * @return array|bool
      */
     public function hGetAllNew($key) {
-        $this->_ext();
-        $this->call();
+        if(!$this->_ext() or !$this->call()) return false;
         $key = $this->getCacheKey($key);
         $keys = $this->handler->hKeys($key);
         $data = [];
         if(!$keys) return $data;
-        $data = $this->handler->hMGet($key, $keys);
-        return $data ? $data : [];
+        $json = $this->handler->hMGet($key, $keys);
+        $data = ($json = is_json($json,true)) ? $json : [];
+        return $data;
     }
 
 //    public function hGetAll_test($key){
@@ -160,11 +166,10 @@ class Redis extends Driver {
     /**
      * @param $name
      * @param $key
-     * @return string
+     * @return string|bool
      */
     public function hGet($name,$key){
-        $this->_ext();
-        $this->call();
+        if(!$this->_ext() or !$this->call()) return false;
         $value = $this->handler->hGet($this->getCacheKey($name),$key);
         $value = ($json = is_json($value,true)) ? $json : $value;
         return $value;
@@ -176,8 +181,7 @@ class Redis extends Driver {
      * @return bool
      */
     public function hSetArray($name,array $array){
-        $this->_ext();
-        $this->call();
+        if(!$this->_ext() or !$this->call()) return false;
         foreach ($array as $key => $value){
             $v = is_scalar($value) ? $value : json_encode($value,JSON_UNESCAPED_UNICODE);
             $this->handler->hSet($this->getCacheKey($name),$key,$v);
@@ -192,8 +196,7 @@ class Redis extends Driver {
      * @return bool|int
      */
     public function hSet($name,$key,$value){
-        $this->_ext();
-        $this->call();
+        if(!$this->_ext() or !$this->call()) return false;
         $v = is_scalar($value) ? $value : json_encode($value,JSON_UNESCAPED_UNICODE);
         return $this->handler->hSet($this->getCacheKey($name),$key,$v);
     }
@@ -204,8 +207,7 @@ class Redis extends Driver {
      * @return bool|int
      */
     public function hDel($name,$keys){
-        $this->_ext();
-        $this->call();
+        if(!$this->_ext() or !$this->call()) return false;
         if(is_array($keys)){
             foreach ($keys as $key){
                 $this->handler->hDel($this->getCacheKey($name),$key);
@@ -224,8 +226,7 @@ class Redis extends Driver {
      * @return mixed
      */
     public function get($name, $default = false) {
-        $this->_ext();
-        $this->call();
+        if(!$this->_ext() or !$this->call()) return false;
         $value = $this->handler->get($this->getCacheKey($name));
         if (is_null($value) || false === $value) {
             return $default;
@@ -247,8 +248,7 @@ class Redis extends Driver {
      * @return boolean
      */
     public function set($name, $value, $expire = null) {
-        $this->_ext();
-        $this->call();
+        if(!$this->_ext() or !$this->call()) return false;
         if (is_null($expire)) {
             $expire = $this->options['expire'];
         }
@@ -277,8 +277,7 @@ class Redis extends Driver {
      * @return false|int
      */
     public function inc($name, $step = 1) {
-        $this->_ext();
-        $this->call();
+        if(!$this->_ext() or !$this->call()) return false;
         $key = $this->getCacheKey($name);
         return $this->handler->incrby($key, $step);
     }
@@ -291,8 +290,7 @@ class Redis extends Driver {
      * @return false|int
      */
     public function dec($name, $step = 1) {
-        $this->_ext();
-        $this->call();
+        if(!$this->_ext() or !$this->call()) return false;
         $key = $this->getCacheKey($name);
         return $this->handler->decrby($key, $step);
     }
@@ -304,8 +302,7 @@ class Redis extends Driver {
      * @return boolean
      */
     public function rm($name) {
-        $this->_ext();
-        $this->call();
+        if(!$this->_ext() or !$this->call()) return false;
         return $this->handler->delete($this->getCacheKey($name));
     }
 
@@ -316,8 +313,7 @@ class Redis extends Driver {
      * @return boolean
      */
     public function clear($tag = null) {
-        $this->_ext();
-        $this->call();
+        if(!$this->_ext() or !$this->call()) return false;
         if ($tag) {
             // 指定标签清除
             $keys = $this->getTagItem($tag);
@@ -332,22 +328,28 @@ class Redis extends Driver {
 
     /**
      * 检查连接
-     * @return bool|string
+     * @param bool $throw
+     * @return bool
      */
-    public function call(){
+    public function call($throw = false){
         try{
             $this->handler->ping();
         }catch (\RedisException $e){
+            log_add($e->getMessage() . ':' . $e->getCode(),'REDIS',__METHOD__);
             if(Tools::isRedisTimeout($e)){
-                if($this->options['persistent']){
+                if($this->options['persistent'] and $this->handler instanceof \Redis){
                     $this->handler->close();
                 }
                 $this->handler = null;
+                $this->is_active = null;
                 self::$_instance = null;
                 self::$_instance = new self();
                 return true;
             }
-            return false;
+            return !$throw ? false : wm_500('redis server timeout');
+        }catch (\Exception $e){
+            log_add($e->getMessage() . ':' . $e->getCode(),'REDIS',__METHOD__);
+            return !$throw ? false : wm_500('redis server timeout');
         }
         return true;
     }
